There was a bug that existed in open GL ES 2.0 implementations and random textures (#239)
* There was a bug that existed in open GL ES 2.0 implementations
where glTexImage2D could not convert GL_RGB formed textures into GL_RGBA
while generating the noise textures.   The GL error
was GL_INVALID_OPERATION(0x502)
This probably made the noise textures blank.
So a new class was added that has the alpha channel for the values.

* Added a fudge for render farming milkdrop where we control the time and the ticks.

* User defined time: Got rid of testing code

* Attempt to fix WIN32 compile - it doesnt use or need the ability to control the time

* Got rid of debug - got rid of minor warning

* projectM.hpp - type in ifdef
projectM_SDL_main.cpp - there is code added to make the initial random number dependent on the time of day, making
the sequence of visualizers more unexpected.  But the posix command time does not exist in Win32 land, so we disable it.