BUILDING.md: upd deps info to universal documentaion
It is impossible to provide&keep the deps list/names for all OSes.

Also different `configure` options require different dependencies provided.

To provide the main documentation on why and what is needed is more important
then a bunch of partial one-liner examples.

& the differences between systems and package names are minimal.

So better describe the required dependencies in terms of Debian/Ubuntu,
and allow maintainers to fiqure-out their OS-specific variants.