BUILDING.md: clean-up & upd docs (#247)
* BUILDING.md: rm tautology & move phrase to relevant place

* BUILDING.md: mv place build info after dependency info & add subsection

`autogen.sh` is in strange state, it consists only of `autoreconf --install`,
and thanks for doing everything right. Also it spews `\n` verbatim,
because person was writing/testing it on mac. Standartizing it in later proposal.

* BUILDING.md: upd deps info to universal documentaion

It is impossible to provide&keep the deps list/names for all OSes.

Also different `configure` options require different dependencies provided.

To provide the main documentation on why and what is needed is more important
then a bunch of partial one-liner examples.

& the differences between systems and package names are minimal.

So better describe the required dependencies in terms of Debian/Ubuntu,
and allow maintainers to fiqure-out their OS-specific variants.

* BUILDING.md: Android: upd commands to be a commands block

* BUILDING.md: mv Windows down, to logically group all *nix info

* BUILDING.md: mv sec Additional info on deps to be after main info on deps

* BUILDING.md: upd text properly outlining command & lib names

* BUILDING.md: Deps: Jack: add mention of libqt5opengl5

Since it uses Qt, so also `libqt5opengl5`.

* BUILDING.md: Additonal info: upd libglm

Co-Authored-By: Mischa Spiegelmock <me@mish.dev>

* BUILDING.md: Building: upd ./configure comments

Co-Authored-By: Mischa Spiegelmock <me@mish.dev>