Fix blank preset bug
I haven't tested this because my only linux box is headless, but this compiles and is the same fix I applied to the Android branch in Feb 2020.  It makes sure than an interrupted peset blend doesn't accidentally leave the master alpha set to a non-1.0 value.  The fix is not very elegant, but it is quite effective.