Fixes issue projectM-visualizer/projectm#130
Adds functionality of writing a rendered frame to file
as requested by the issue https://github.com/projectM-visualizer/projectm/issues/130.

Keyhandler.cpp has been updated to handle the keydown of 'd',
which triggers writeNextFrameToFile to be set to true, which means
at the end of Renderer::FinishPass1 we write the texture contents to file.

Renderer::debugWriteMainTextureToFile() const contains the functionality
of writing the contents of the main texture.