Simplify MasterRenderItemDistance::ComputeDistance
This should in practice do more or less what it already did previously.
I think it was written like this to allow easily add on new ways of
merging.
This was not used, but added a lot of complexity.
Better simplify things and then add this ability if/when need comes.