Avoid relying on max() from windows.h
Unfortunately, windows.h breaks the standard library because it
defines max() as a macro by default.

To fix this, add NOMINMAX (and WIN32_LEAN_AND_MEAN for good measure) to
the compile options. Ideally, this would be set on a per-target basis,
but with so many OBJECT targets, that is really tedious. Luckily,
add_compile_definitions() is scoped to the current directory and sub-
directories only.